third_party/WebKit/LayoutTests/webaudio/audiocontext-suspend-resume-expected.txt

Test suspend/resume for an AudioContext

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof context.suspend() is equal to function.
PASS typeof context.resume() is equal to function.
PASS context.suspend() resolved correctly.
PASS Calling multiple context.suspend() did not throw an exception.
PASS Calling context.suspend() after close() rejected correctly (with InvalidStateError: Cannot suspend a context that has been closed).
PASS Calling context.resume() after close() rejected correctly (with InvalidAccessError: cannot resume a closed AudioContext).
PASS successfullyParsed is true

TEST COMPLETE
